requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
loguru==0.7.0
multidict==6.0.4
numpy==1.24.3
openai==0.27.6
pinecone-client==2.2.1
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
yarl==1.9.2
